requirements.txt
langchain==1.0.5
langchain-openai==1.0.2
langchain-community==0.4.1
faiss-cpu==1.8.0.post1
pypdf==5.0.1
tiktoken==0.7.0
python-dotenv==1.1.1
openai==2.6.1
langfuse>=3.0.0
rich==13.8.1
pydantic==2.9.2